// src/generated/models/EnrichmentRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * XYO RESTful Transaction Enrichment API
 * RESTful API for enriching financial transactions.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface EnrichmentRequest
 */
export interface EnrichmentRequest {
  /**
   * Content is the payment description, maximum 128 characters.
   * @type {string}
   * @memberof EnrichmentRequest
   */
  content: string
  /**
   * CountryCode is the ISO 3166-1 alpha-2 two-character country code.
   * @type {string}
   * @memberof EnrichmentRequest
   */
  countryCode: string
}

/**
 * Check if a given object implements the EnrichmentRequest interface.
 */
export function instanceOfEnrichmentRequest(
  value: object,
): value is EnrichmentRequest {
  if (
    !('content' in value) ||
    value['content'] === undefined
  )
    return false
  if (
    !('countryCode' in value) ||
    value['countryCode'] === undefined
  )
    return false
  return true
}

export function EnrichmentRequestFromJSON(
  json: any,
): EnrichmentRequest {
  return EnrichmentRequestFromJSONTyped(json, false)
}

export function EnrichmentRequestFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): EnrichmentRequest {
  if (json == null) {
    return json
  }
  return {
    content: json['content'],
    countryCode: json['countryCode'],
  }
}

export function EnrichmentRequestToJSON(
  json: any,
): EnrichmentRequest {
  return EnrichmentRequestToJSONTyped(json, false)
}

export function EnrichmentRequestToJSONTyped(
  value?: EnrichmentRequest | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    content: value['content'],
    countryCode: value['countryCode'],
  }
}