// src/generated/models/EnrichTransactionsRequestInner.ts
/* tslint:disable */
/* eslint-disable */
/**
 * XYO RESTful Transaction Enrichment API
 * RESTful API for enriching financial transactions.
 *
 * The version of the OpenAPI document: 2.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EnrichTransactionsRequestInner
 */
export interface EnrichTransactionsRequestInner {
    /**
     * Content is the payment description, maximum 128 characters.
     * @type {string}
     * @memberof EnrichTransactionsRequestInner
     */
    content: string;
    /**
     * CountryCode is the ISO 3166-1 alpha-2 two-character country code.
     * @type {string}
     * @memberof EnrichTransactionsRequestInner
     */
    countryCode: string;
}

/**
 * Check if a given object implements the EnrichTransactionsRequestInner interface.
 */
export function instanceOfEnrichTransactionsRequestInner(value: object): value is EnrichTransactionsRequestInner {
    if (!('content' in value) || value['content'] === undefined) return false;
    if (!('countryCode' in value) || value['countryCode'] === undefined) return false;
    return true;
}

export function EnrichTransactionsRequestInnerFromJSON(json: any): EnrichTransactionsRequestInner {
    return EnrichTransactionsRequestInnerFromJSONTyped(json, false);
}

export function EnrichTransactionsRequestInnerFromJSONTyped(json: any, ignoreDiscriminator: boolean): EnrichTransactionsRequestInner {
    if (json == null) {
        return json;
    }
    return {
        
        'content': json['content'],
        'countryCode': json['countryCode'],
    };
}

export function EnrichTransactionsRequestInnerToJSON(json: any): EnrichTransactionsRequestInner {
    return EnrichTransactionsRequestInnerToJSONTyped(json, false);
}

export function EnrichTransactionsRequestInnerToJSONTyped(value?: EnrichTransactionsRequestInner | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'content': value['content'],
        'countryCode': value['countryCode'],
    };
}