// src/generated/models/EnrichmentCollectionStatusResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * XYO RESTful Transaction Enrichment API
 * RESTful API for enriching financial transactions.
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime'
/**
 *
 * @export
 * @interface EnrichmentCollectionStatusResponse
 */
export interface EnrichmentCollectionStatusResponse {
  /**
   * Status represents the processing state of a bulk enrichment job.
   * @type {EnrichmentCollectionStatusResponseStatusEnum}
   * @memberof EnrichmentCollectionStatusResponse
   */
  status: EnrichmentCollectionStatusResponseStatusEnum
}

/**
 * @export
 */
export const EnrichmentCollectionStatusResponseStatusEnum =
  {
    Ready: 'READY',
    Failed: 'FAILED',
    Pending: 'PENDING',
  } as const
export type EnrichmentCollectionStatusResponseStatusEnum =
  (typeof EnrichmentCollectionStatusResponseStatusEnum)[keyof typeof EnrichmentCollectionStatusResponseStatusEnum]

/**
 * Check if a given object implements the EnrichmentCollectionStatusResponse interface.
 */
export function instanceOfEnrichmentCollectionStatusResponse(
  value: object,
): value is EnrichmentCollectionStatusResponse {
  if (!('status' in value) || value['status'] === undefined)
    return false
  return true
}

export function EnrichmentCollectionStatusResponseFromJSON(
  json: any,
): EnrichmentCollectionStatusResponse {
  return EnrichmentCollectionStatusResponseFromJSONTyped(
    json,
    false,
  )
}

export function EnrichmentCollectionStatusResponseFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): EnrichmentCollectionStatusResponse {
  if (json == null) {
    return json
  }
  return {
    status: json['status'],
  }
}

export function EnrichmentCollectionStatusResponseToJSON(
  json: any,
): EnrichmentCollectionStatusResponse {
  return EnrichmentCollectionStatusResponseToJSONTyped(
    json,
    false,
  )
}

export function EnrichmentCollectionStatusResponseToJSONTyped(
  value?: EnrichmentCollectionStatusResponse | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value
  }

  return {
    status: value['status'],
  }
}